+++ src/components/functions/get-review-route.ts
@@ -1,0 +1,5 @@
+import { AppRoute } from '../app';
+
+export function getReviewRoute(filmId: string): string {
+  return AppRoute.ReviewPage.replace(':id', `:id=${filmId}`);
+}
