+++ src/components/tab-details.tsx
