+++ src/store/index.ts
@@ -2,9 +2,9 @@ import { configureStore } from '@reduxjs/toolkit';
 import { appReducer } from './reducer';
 
 const store = configureStore({
- reducer: {
-   app: appReducer,
- },
+  reducer: {
+    app: appReducer,
+  },
 });
 
 export default store;
