+++ src/store/reducer.ts
@@ -1,5 +1,5 @@
-import { changeGenre, setFilms } from "./action";
-import { Film } from "../mocks/films";
+import { changeGenre, setFilms } from './action';
+import { Film } from '../mocks/films';
 import { createReducer } from '@reduxjs/toolkit';
 
 type AppState = {
@@ -8,16 +8,16 @@ type AppState = {
 };
 
 const initialState: AppState = {
- genre: '',
- films: [],
+  genre: '',
+  films: [],
 };
 
 export const appReducer = createReducer(initialState, (builder) => {
- builder
-   .addCase(changeGenre, (state, action) => {
-     state.genre = action.payload;
-   })
-   .addCase(setFilms, (state, action) => {
-     state.films = action.payload;
-   });
+  builder
+    .addCase(changeGenre, (state, action) => {
+      state.genre = action.payload;
+    })
+    .addCase(setFilms, (state, action) => {
+      state.films = action.payload;
+    });
 });
